Moved mainloop into Snake_Main file.
The idea is for the "Snake_Main.py" file to house all of the core game logic whilst the "Snake_Display.py" file exists for graphical changes.
This means that while something like the "makeGrid" function would belong to "Snake_Display.py" the mainloop should belong to "Snake_Main.py".