"displaySnake" no longer uses currentNode.next
Previously "displaySnake" would only try to display a 'snakeNode' if that node had a node after it (or was the head). After this it would try to display the current node. This means that it would display the head, enter the loop, then display the head again before ending.
I don't know why this worked with more than 2 but honestly I don't care to think about it right now.