Changed "displayApple" parameters
I realized that I can not just check the color of a rect in pygame. Because of this I need to store the actual location of the apple in a variable and it makes most sense to have that be a list of coordinates.
Now that the data is stored in this way it's makes sense for the corresponding function to be responsible for unpacking that data.